Improve version resolution with osxcross-macports
I rejigged the package resolution logic to allow for older versions to be downloaded when the latest advertised one is not available for the target being requested.

When required and a fat binary is all that is available which contains the arch in question (e.g. a universal build contains arm64 but there is no arm64 package available) we will grab that.

Due to the change for macOS 11+ to use the arm64 packages by default, I added a flag to download x86_64 versions of packages so people building for x86 macs still can.

macports currently doesn't build arm64e universal packages, so when searching for newer darwin universal binaries it will fail, I removed arm64e from the list of arches.

This makes the wrapper a bit more resilient for versions of Darwin where builds may be missing for a package version.